[clean] Trying to simplify higher-level, global interfaces
